Refactoring order book, cambio en el diseño y manejo de archivos
